src/requirements.txt
PyMuPDF==1.22.5
python-dotenv==1.0.0